refactor: replacing chai and nyc with node test runner * removed nyc settings from package.json * updated npm run for test and coverage * updated all tests * rewrote some tests to be one assert per test * removed unused chai and mocha dependencies (module count from 472 to 293)
