+++ package.json
@@ -15,46 +15,23 @@
     "yargs": "^17.7.2"
   },
   "devDependencies": {
-    "mocha": "^10.2.0",
-    "@istanbuljs/nyc-config-typescript": "^1.0.2",
     "@octokit/types": "^11.1.0",
-    "@types/chai": "^4.3.5",
     "@types/cli-progress": "^3.11.0",
-    "@types/mocha": "^10.0.1",
     "@types/node": "^20.5.0",
     "@types/sinon": "^10.0.16",
     "@types/yargs": "^17.0.24",
     "@typescript-eslint/eslint-plugin": "^6.3.0",
-    "chai": "^4.3.7",
     "eslint": "8.47.0",
-    "nyc": "^15.1.0",
     "sinon": "^15.2.0",
-    "ts-mocha": "^10.0.0",
     "ts-node": "^10.9.1",
     "tsx": "^3.12.7"
   },
   "optionalDependencies": {
     "fsevents": "2.3.2"
   },
-  "nyc": {
-    "extends": "@istanbuljs/nyc-config-typescript",
-    "check-coverage": true,
-    "all": true,
-    "include": [
-      "src/**/!(*.test.*).[tj]s?(x)"
-    ],
-    "reporter": [
-      "html",
-      "lcov",
-      "text",
-      "text-summary"
-    ],
-    "report-dir": "coverage"
-  },
   "scripts": {
-    "test": "ts-mocha -n es-module-specifier-resolution=node -n loader=ts-node/esm/transpile-only -p tsconfig.json 'src/**/*.test.ts'",
-    "test:coverage": "nyc npm run test",
-    "nodetest": "node --experimental-test-coverage --loader tsx --test src/**/*.nodetest.ts  src/**/**/*.nodetest.ts",
+    "test:coverage": "node --experimental-test-coverage --loader tsx --test src/**/*.test.ts  src/**/**/*.test.ts",
+    "test": "node --loader tsx --test  src/**/**/*.test.ts",
     "start": "node --no-warnings --es-module-specifier-resolution=node --loader ts-node/esm/transpile-only src/Main.ts",
     "license_checker": "node --no-warnings --es-module-specifier-resolution=node --loader ts-node/esm/transpile-only src/Main.ts"
   }
