+++ src/Domain/Downloader.test.ts
@@ -1,17 +1,16 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import 'mocha';
-import {stub, restore} from 'sinon';
-import {assert} from 'chai';
-import {Downloader} from './Downloader';
-import {GithubClient} from '../Adapter/Import/GithubClient';
-import {HTTPClient} from '../Adapter/Import/HTTPClient';
+import assert from 'node:assert';
+import { afterEach, beforeEach, describe, it } from 'node:test';
+import { stub, restore } from 'sinon';
+import { Downloader } from './Downloader';
+import { GithubClient } from '../Adapter/Import/GithubClient';
+import {HTTPClient } from '../Adapter/Import/HTTPClient';
 import {Logger} from '../Logging/Logging';
 
-
-describe('downloadLicenseFromGithub', function () {
+describe('downloadLicenseFromGithub', () => {
     Logger.setSilent();
     let downloader: Downloader;
-    this.beforeEach(function () {
+    
+    beforeEach(()=>{
         downloader = new Downloader();
         const downloadRepoStub = stub(GithubClient.prototype, 'downloadRepo');
         const httpStub = stub(HTTPClient.prototype, 'getWebsite');
@@ -41,8 +40,8 @@ describe('downloadLicenseFromGithub', function () {
             }));
 
         downloadRepoStub.onCall(0).returns(
-            new Promise<any>((resolve) => {
-                const value: any = [
+            new Promise<unknown>((resolve) => {
+                const value: unknown = [
                     {
                         'name': 'license',
                         'download_url': 'license url'
@@ -58,8 +57,8 @@ describe('downloadLicenseFromGithub', function () {
         );
 
         downloadRepoStub.onCall(1).returns(
-            new Promise<any>((resolve) => {
-                const value: any = [
+            new Promise<unknown>((resolve) => {
+                const value: unknown = [
                     {
                         'name': 'license.md',
                         'download_url': 'license.md url'
@@ -72,23 +71,23 @@ describe('downloadLicenseFromGithub', function () {
                 resolve(value);
             })
         );
-
     });
-
-    this.afterEach(() => {
-        restore();
-    });
+    
     it('should download the license files if it exists in the given repo', async function () {
         let res = await downloader.downloadDataFromGithub('github.com/test/repo');
         assert.deepEqual(res, ['license', 'readme']);
         res = await downloader.downloadDataFromGithub('github.com/test/repo');
         assert.deepEqual(res, ['license', 'readme']);
     });
+
+    afterEach(()=>{
+        restore();
+    });
 });
 
 describe('filterRepoInfoFromURL', function () {
     let downloader: Downloader;
-    this.beforeEach(function () {
+    beforeEach(() => {
         downloader = new Downloader();
     });
     it('should correctly extract the user and repository from the given url', function () {
